+++ front_end/src/containers/CardDeck/CardDeck.js
@@ -2,8 +2,8 @@ import React, { Fragment } from 'react';
 import axios from 'axios';
 
 import store from '../../store';
-import PaginateButtons from '../UI/PaginateButtons/PaginateButtons';
-import Card from './Card/Card';
+import PaginateButtons from '../../components/UI/PaginateButtons/PaginateButtons';
+import Card from '../../components/Card/Card';
 import CSS from './CardDeck.module.css';
